get it to at least compile

+++ src/Amazon.S3/Model/S3BucketLifecycleConfigurationTransition.m
@@ -32,7 +32,7 @@
     [xml appendString:@"<Transition>"];
     [xml appendFormat:@"<StorageClass>%@</StorageClass>", self.storageClass];
     if (self.transitionDays > 0) {
-        [xml appendFormat:@"<Days>%d</Days>", self.transitionDays];
+        [xml appendFormat:@"<Days>%ld</Days>", (long)self.transitionDays];
     }
     else if (self.transitionDate != nil) {
         [xml appendFormat:@"<Date>%@</Date>", [self.transitionDate stringWithISO8061Format]];
